requirements.txt
certifi==2024.2.2
python-telegram-bot==20.8
aiohttp==3.9.1
sqlalchemy==2.0.25
pandas==2.1.4
numpy==1.26.3
python-dotenv==1.0.0
scikit-learn==1.4.0